requirements.txt
streamlit
crewai
python-dotenv
psycopg2-binary
SQLAlchemy
PyYAML
openai
